Change base type of enums to byte task: https://github.com/MorphineForOdin/GoT2_LCG/issues/30

+++ GotLcg/GotLcg.Database.SqlServer/Data/BootstrapData/DATA_CardType.sql
@@ -2,7 +2,7 @@
 
 DECLARE @cardTypeData AS TABLE
 (
-	[Id] INT NOT NULL,
+	[Id] TINYINT NOT NULL,
 	[Name] NVARCHAR(10) NOT NULL
 );
