+++ GotLcg/GotLcg.Database.SqlServer/Tables/ChallengeIcon.sql
@@ -1,8 +1,8 @@
 ﻿CREATE TABLE [dbo].[ChallengeIcon]
 (
-	[Id] INT NOT NULL, 
+	[Id] TINYINT NOT NULL,
     [Name] NVARCHAR(10) NOT NULL,
-    [CreatedDate] DATETIME2 NOT NULL CONSTRAINT [DC_ChallengeIcon_CreatedDate] DEFAULT GETUTCDATE(), 
+    [CreatedDate] DATETIME2 NOT NULL CONSTRAINT [DC_ChallengeIcon_CreatedDate] DEFAULT GETUTCDATE(),
     [ModifiedDate] DATETIME2 NOT NULL CONSTRAINT [DC_ChallengeIcon_ModifiedDate] DEFAULT GETUTCDATE(),
 	CONSTRAINT [PK_ChallengeIcon] PRIMARY KEY CLUSTERED ([Id] ASC)
 );
